scraper/pyproject.toml
[project]
name = "istanbulmedic-scraper"
version = "0.1.0"
description = "Doctor credential scraper for ISHRS and IAHRS public directories."
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "python-dotenv>=1.0",
  "supabase>=2.7",
  "unidecode>=1.3",
  "rapidfuzz>=3.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.14",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
